refactor(api-client): fixed method name
