+++ packages/api-client/src/api/index.ts
@@ -4,7 +4,7 @@ export { default as addProductsToCart } from './addProductsToCart';
 export { default as addProductToWishList } from './addProductToWishList';
 export { default as addSimpleProductsToCart } from './addSimpleProductsToCart';
 export { default as applyCouponToCart } from './applyCouponToCart';
-export { default as bundledProductDetail } from './bundledProductDetail';
+export { default as bundleProductDetail } from './bundleProductDetail';
 export { default as cart } from './cart';
 export { default as categoryList } from './categoryList';
 export { default as categorySearch } from './categorySearch';
