+++ packages/api-client/src/index.ts
@@ -50,8 +50,8 @@ export {
   Breadcrumb,
   BundleCartItem,
   BundleCreditMemoItem,
-  BundledProductDetailQuery,
-  BundledProductDetailQueryVariables,
+  BundleProductDetailQuery,
+  BundleProductDetailQueryVariables,
   BundleInvoiceItem,
   BundleItem,
   BundleItemOption,
