+++ packages/api-client/src/types/GraphQL.ts
@@ -6647,7 +6647,7 @@ export type ApplyCouponToCartMutationVariables = Exact<{
 
 export type ApplyCouponToCartMutation = { applyCouponToCart?: Maybe<{ cart: { id: string, email?: Maybe<string>, is_virtual: boolean, total_quantity: number, applied_coupons?: Maybe<Array<Maybe<{ code: string }>>>, prices?: Maybe<{ subtotal_excluding_tax?: Maybe<{ value?: Maybe<number> }>, subtotal_including_tax?: Maybe<{ value?: Maybe<number> }>, applied_taxes?: Maybe<Array<Maybe<{ label: string, amount: { value?: Maybe<number> } }>>>, discounts?: Maybe<Array<Maybe<{ label: string, amount: { value?: Maybe<number> } }>>>, grand_total?: Maybe<{ value?: Maybe<number> }> }>, items?: Maybe<Array<Maybe<{ uid: string, quantity: number, bundle_options: Array<Maybe<{ uid: string, label: string, type: string, values: Array<Maybe<{ id: number, label: string, price: number, quantity: number }>> }>>, product: { __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>>, configurable_options?: Maybe<Array<Maybe<{ attribute_code?: Maybe<string>, attribute_uid: string, label?: Maybe<string>, position?: Maybe<number>, uid: string, use_default?: Maybe<boolean>, values?: Maybe<Array<Maybe<{ label?: Maybe<string>, uid?: Maybe<string>, swatch_data?: Maybe<{ value?: Maybe<string> } | { value?: Maybe<string> } | { value?: Maybe<string> }> }>>> }>>> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> }, prices?: Maybe<{ row_total: { value?: Maybe<number> }, row_total_including_tax: { value?: Maybe<number> }, total_item_discount?: Maybe<{ value?: Maybe<number> }> }> } | { uid: string, quantity: number, configurable_options: Array<Maybe<{ configurable_product_option_uid: string, option_label: string, configurable_product_option_value_uid: string, value_label: string }>>, product: { __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>>, configurable_options?: Maybe<Array<Maybe<{ attribute_code?: Maybe<string>, attribute_uid: string, label?: Maybe<string>, position?: Maybe<number>, uid: string, use_default?: Maybe<boolean>, values?: Maybe<Array<Maybe<{ label?: Maybe<string>, uid?: Maybe<string>, swatch_data?: Maybe<{ value?: Maybe<string> } | { value?: Maybe<string> } | { value?: Maybe<string> }> }>>> }>>> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> }, prices?: Maybe<{ row_total: { value?: Maybe<number> }, row_total_including_tax: { value?: Maybe<number> }, total_item_discount?: Maybe<{ value?: Maybe<number> }> }> } | { uid: string, quantity: number, product: { __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>>, configurable_options?: Maybe<Array<Maybe<{ attribute_code?: Maybe<string>, attribute_uid: string, label?: Maybe<string>, position?: Maybe<number>, uid: string, use_default?: Maybe<boolean>, values?: Maybe<Array<Maybe<{ label?: Maybe<string>, uid?: Maybe<string>, swatch_data?: Maybe<{ value?: Maybe<string> } | { value?: Maybe<string> } | { value?: Maybe<string> }> }>>> }>>> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> }, prices?: Maybe<{ row_total: { value?: Maybe<number> }, row_total_including_tax: { value?: Maybe<number> }, total_item_discount?: Maybe<{ value?: Maybe<number> }> }> } | { uid: string, quantity: number, product: { __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>>, configurable_options?: Maybe<Array<Maybe<{ attribute_code?: Maybe<string>, attribute_uid: string, label?: Maybe<string>, position?: Maybe<number>, uid: string, use_default?: Maybe<boolean>, values?: Maybe<Array<Maybe<{ label?: Maybe<string>, uid?: Maybe<string>, swatch_data?: Maybe<{ value?: Maybe<string> } | { value?: Maybe<string> } | { value?: Maybe<string> }> }>>> }>>> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> }, prices?: Maybe<{ row_total: { value?: Maybe<number> }, row_total_including_tax: { value?: Maybe<number> }, total_item_discount?: Maybe<{ value?: Maybe<number> }> }> } | { uid: string, quantity: number, product: { __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>>, configurable_options?: Maybe<Array<Maybe<{ attribute_code?: Maybe<string>, attribute_uid: string, label?: Maybe<string>, position?: Maybe<number>, uid: string, use_default?: Maybe<boolean>, values?: Maybe<Array<Maybe<{ label?: Maybe<string>, uid?: Maybe<string>, swatch_data?: Maybe<{ value?: Maybe<string> } | { value?: Maybe<string> } | { value?: Maybe<string> }> }>>> }>>> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, rating_summary: number, review_count: number, url_key?: Maybe<string>, reviews: { items: Array<Maybe<{ average_rating: number, ratings_breakdown: Array<Maybe<{ name: string, value: string }>> }>> }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, url_rewrites?: Maybe<Array<Maybe<{ url?: Maybe<string> }>>>, categories?: Maybe<Array<Maybe<{ uid: string, name?: Maybe<string>, url_suffix?: Maybe<string>, url_path?: Maybe<string>, breadcrumbs?: Maybe<Array<Maybe<{ category_name?: Maybe<string>, category_url_path?: Maybe<string> }>>> }>>> }, prices?: Maybe<{ row_total: { value?: Maybe<number> }, row_total_including_tax: { value?: Maybe<number> }, total_item_discount?: Maybe<{ value?: Maybe<number> }> }> }>>>, shipping_addresses: Array<Maybe<{ firstname: string, lastname: string, street: Array<Maybe<string>>, city: string, company?: Maybe<string>, postcode?: Maybe<string>, telephone: string, selected_shipping_method?: Maybe<{ carrier_code: string, carrier_title: string, method_code: string, method_title: string, amount: { value?: Maybe<number>, currency?: Maybe<CurrencyEnum> } }>, region?: Maybe<{ code?: Maybe<string>, region_id?: Maybe<number>, label?: Maybe<string> }>, country: { code: string, label: string } }>>, billing_address?: Maybe<{ firstname: string, lastname: string, street: Array<Maybe<string>>, city: string, company?: Maybe<string>, postcode?: Maybe<string>, telephone: string, region?: Maybe<{ code?: Maybe<string>, region_id?: Maybe<number>, label?: Maybe<string> }>, country: { code: string, label: string } }> } }> };
 
-export type BundledProductDetailQueryVariables = Exact<{
+export type BundleProductDetailQueryVariables = Exact<{
   search?: Maybe<Scalars['String']>;
   filter?: Maybe<ProductAttributeFilterInput>;
   pageSize?: Maybe<Scalars['Int']>;
@@ -6656,7 +6656,7 @@ export type BundledProductDetailQueryVariables = Exact<{
 }>;
 
 
-export type BundledProductDetailQuery = { products?: Maybe<{ items?: Maybe<Array<Maybe<{ uid: string, items?: Maybe<Array<Maybe<{ __typename: 'BundleItem', position?: Maybe<number>, required?: Maybe<boolean>, sku?: Maybe<string>, title?: Maybe<string>, type?: Maybe<string>, uid?: Maybe<string>, options?: Maybe<Array<Maybe<{ __typename: 'BundleItemOption', can_change_quantity?: Maybe<boolean>, is_default?: Maybe<boolean>, label?: Maybe<string>, position?: Maybe<number>, price?: Maybe<number>, price_type?: Maybe<PriceTypeEnum>, quantity?: Maybe<number>, uid: string, product?: Maybe<{ __typename: 'BundleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> } | { __typename: 'ConfigurableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> } | { __typename: 'DownloadableProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> } | { __typename: 'GroupedProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> } | { __typename: 'SimpleProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> } | { __typename: 'VirtualProduct', uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } }, thumbnail?: Maybe<{ url?: Maybe<string>, position?: Maybe<number>, disabled?: Maybe<boolean>, label?: Maybe<string> }> }> }>>> }>>> } | { uid: string } | { uid: string } | { uid: string } | { uid: string } | { uid: string }>>> }> };
+export type BundleProductDetailQuery = { products?: Maybe<{ items?: Maybe<Array<Maybe<{ uid: string, items?: Maybe<Array<Maybe<{ __typename: 'BundleItem', position?: Maybe<number>, required?: Maybe<boolean>, sku?: Maybe<string>, title?: Maybe<string>, type?: Maybe<string>, uid?: Maybe<string>, options?: Maybe<Array<Maybe<{ __typename: 'BundleItemOption', can_change_quantity?: Maybe<boolean>, is_default?: Maybe<boolean>, position?: Maybe<number>, uid: string, product?: Maybe<{ uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } } | { uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } } | { uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } } | { uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } } | { uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } } | { uid: string, sku?: Maybe<string>, name?: Maybe<string>, stock_status?: Maybe<ProductStockStatus>, only_x_left_in_stock?: Maybe<number>, price_range: { maximum_price?: Maybe<{ final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } }>, minimum_price: { final_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> }, regular_price: { currency?: Maybe<CurrencyEnum>, value?: Maybe<number> } } } }> }>>> }>>> } | { uid: string } | { uid: string } | { uid: string } | { uid: string } | { uid: string }>>> }> };
 
 export type CartQueryVariables = Exact<{
   cartId: Scalars['String'];
